+++ src/components/button-more/button-more.html
@@ -1,4 +1,7 @@
-<md-button class="button-more md-icon-button" aria-label="More" ng-click="$mdMenu.open($event)">
-  <md-icon md-font-set="material-icons" class="button-more__more-icon ">more_vert</md-icon>
-</md-button>
-<popup></popup>
+<md-menu>
+  <md-button aria-label="Open menu with custom trigger" class="button-more md-icon-button" aria-label="More" ng-click="$mdMenu.open()">
+    <md-icon md-font-set="material-icons" class="button-more__more-icon ">more_vert</md-icon>
+  </md-button>
+  <popup></popup>
+  <!--<popup popup-opened="$ctrl.popupOpened"></popup>-->
+</md-menu>
