+++ src/components/button-more/button-more.js
@@ -3,10 +3,17 @@ import "angular-animate";
 import "angular-aria";
 import "angular-material";
 import template from "./button-more.html";
-import { popupModule } from "../popup/popup";
-class ButtonMoreController {}
+import { popupModule } from "./popup/popup";
+class ButtonMoreController {
+  // constructor(){
+  //   this.popupOpened  = false;
+  // }
+  // open(event){
+  //   this.popupOpened = !this.popupOpened;
+  // }
+}
 export const buttonMoreModule = angular
-  .module("buttonMore", popupModule, ["ngMaterial"])
+  .module("buttonMore", ["ngMaterial", popupModule])
   .component("buttonMore", {
     template,
     controller: ButtonMoreController
