+++ src/components/button-more/popup/popup.html
@@ -1,0 +1,5 @@
+<md-menu-content class="popup-menu" width="4" ng-click="$mdMenu.close()">
+  <md-menu-item class="popup-menu__item" ng-repeat="item in [1, 2, 3]" z-index="9999999999999">
+
+  </md-menu-item>
+</md-menu-content>
