--- src/components/popup/popup.html
@@ -1,7 +1,0 @@
-<md-menu md-offset="0 -7">
-  <md-menu-content width="6">
-    <md-menu-item ng-repeat="item in [1, 2, 3]">
-      <md-button ng-click="ctrl.announceClick($index)"> <span md-menu-align-target="">Option</span> {{item}} </md-button>
-    </md-menu-item>
-  </md-menu-content>
-</md-menu>
