+++ src/index.scss
@@ -7,5 +7,5 @@
 @import "./components/button-more/button-more";
 @import "./components/drawer-menu/drawer-menu";
 @import "./components/tags/tags";
-@import "./components/popup/popup";
+@import "./components/button-more/popup/popup";
 @import "./components/app/app";
